+++ RELEASE_NOTES.md
@@ -132,6 +132,7 @@ This serves two purposes:
   - This also removes the `<x-hyde::docs.search-input />` and `<x-hyde::docs.search-scripts />` Blade components, replaced by the new `<x-hyde::docs.hyde-search />` component.
 - Removed the `.torchlight-enabled` CSS class in https://github.com/hydephp/develop/pull/2036.
 - Removed The `hyde.css` file from HydeFront in https://github.com/hydephp/develop/pull/2037 as all styles were refactored to Tailwind in https://github.com/hydephp/develop/pull/2024.
+- Removed the `MarkdownService::withPermalinks` method in https://github.com/hydephp/develop/pull/2047
 
 ### Fixed